Electricity billing system
The project is a program to calculate the total bill of a customer for the C & C Limited is an electricity distribution company with retail tariff rates applicable for their Domestic and Commercial customers categories for the year 2024. It considers factors such as age for discounts and different unit rates for peak, shoulder, and off-peak consumption.